+++ src/org/openscience/jmol/app/jmolpanel/AtomSetChooser.java
@@ -135,28 +135,28 @@ ActionListener, ChangeListener, Runnable {
   /**
    * Precision of the vibration scale slider
    */
-  private static final float AMPLITUDE_PRECISION = 0.01f;
+  private static final double AMPLITUDE_PRECISION = 0.01f;
   /**
    * Maximum value for vibration scale. Should be in preferences?
    */
-  private static final float AMPLITUDE_MAX = 1;
+  private static final double AMPLITUDE_MAX = 1;
   /**
    * Initial value of vibration scale. Should be in preferences?
    */
-  private static final float AMPLITUDE_VALUE = 0.5f;
+  private static final double AMPLITUDE_VALUE = 0.5d;
 
   /**
    * Precision of the vibration period slider in seconds.
    */
-  private static final float PERIOD_PRECISION = 0.001f;
+  private static final double PERIOD_PRECISION = 0.001f;
   /**
    * Maximum value for the vibration period in seconds. Should be in preferences?
    */
-  private static final float PERIOD_MAX = 1; // in seconds
+  private static final double PERIOD_MAX = 1; // in seconds
   /**
    * Initial value for the vibration period in seconds. Should be in preferences?
    */
-  private static final float PERIOD_VALUE = 0.5f;
+  private static final double PERIOD_VALUE = 0.5d;
 
   /**
    * Maximum value for vector radius.
@@ -170,15 +170,15 @@ ActionListener, ChangeListener, Runnable {
   /**
    * Precision of the vector scale slider
    */
-  private static final float SCALE_PRECISION = 0.01f;
+  private static final double SCALE_PRECISION = 0.01f;
   /**
    * Maximum value for vector scale. Should be in preferences?
    */
-  private static final float SCALE_MAX = 2.0f;
+  private static final double SCALE_MAX = 2.0d;
   /**
    * Initial value of vector scale. Should be in preferences?
    */
-  private static final float SCALE_VALUE = 1.0f;
+  private static final double SCALE_VALUE = 1.0d;
